projects/simple_chatbot_youtube/pyproject.toml
[tool.poetry]
name = "simple-chatbot"
version = "0.1.0"
description = ""
authors = ["Minh-Duc Bui <minhduc2k31905@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.51.2"
python-dotenv = "^1.0.1"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
streamlit = "^1.39.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
